# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MSS Platform API
  description: |
    MSS Platform API provides comprehensive endpoints for managing cybersecurity services,
    multi-tenant organizations, user management, and security operations.
    
    ## Authentication
    
    The API uses JWT Bearer tokens for authentication. Most endpoints require a valid JWT token
    in the Authorization header:
    
    ```
    Authorization: Bearer <jwt_token>
    ```
    
    ## Multi-Factor Authentication Flow
    
    1. **Login**: POST /api/auth/login with credentials
       - Returns `mfa_token` if MFA is enabled
       - Returns `access_token` and `refresh_token` if MFA is disabled
    
    2. **Complete MFA**: POST /api/auth/mfa/complete with MFA token and TOTP/backup code
       - Returns `access_token` and `refresh_token`
    
    3. **Refresh Token**: POST /api/auth/refresh when access token expires
    
    ## Role-Based Access Control
    
    ### Customer Organization Roles
    - **admin**: Full organization management
    - **report_viewer**: Read access to security reports
    - **request_user**: Can request interventions
    - **basic_user**: Limited dashboard access
    
    ### MSS Provider Roles
    - **super_admin**: Platform-wide administration
    - **account_manager**: Customer organization management
    - **security_analyst**: Security monitoring and analysis
    - **technician**: Customer environment access
    
    ## Rate Limiting
    
...
    
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from login/MFA completion
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message
        error:
          type: string
          description: Error message if success is false
      required:
        - success
...
    SecurityReport:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        type:
          type: string
          enum: [vulnerability_assessment, penetration_test, compliance_audit, incident_response, monthly_summary]
        status:
          type: string
          enum: [draft, pending_review, approved, published]
        organizationId:
          type: string
          format: uuid
        generatedBy:
          type: string
          format: uuid
        approvedBy:
          type: string
          format: uuid
          nullable: true
        summary:
          type: string
          description: Executive summary
        findings:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              title:
                type: string
              severity:
                type: string
                enum: [low, medium, high, critical]
              description:
                type: string
              recommendation:
                type: string
              status:
                type: string
                enum: [open, in_progress, resolved]
        reportPeriod:
          type: object
          properties:
            startDate:
              type: string
              format: date
            endDate:
              type: string
              format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - title
        - type
        - status
        - organizationId
        - generatedBy
        - createdAt
        - updatedAt

    # Intervention schemas
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: false
            error:
              type: string
              description: Error message
            code:
              type: string
              description: Error code for programmatic handling
              example: VALIDATION_ERROR
            details:
              type: object
              additionalProperties: true
              description: Additional error details
          required:
            - success
            - error
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Authentication required"
            code: "UNAUTHORIZED"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Insufficient permissions"
            code: "FORBIDDEN"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Resource not found"
            code: "NOT_FOUND"
...
paths:
...
  /reports:
    get:
      tags:
        - Security Reports
      summary: Get security reports
      description: |
        Get security reports based on user permissions:
        - Customer users: See reports for their organization
        - MSS provider users: See reports for all customer organizations
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: type
          in: query
          schema:
            type: string
            enum: [vulnerability_assessment, penetration_test, compliance_audit, incident_response, monthly_summary]
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending_review, approved, published]
        - name: orgId
          in: query
          schema:
            type: string
            format: uuid
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Reports retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          reports:
                            type: array
                            items:
                              $ref: '#/components/schemas/SecurityReport'
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalPages:
                            type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /reports/{reportId}:
    get:
      tags:
        - Security Reports
      summary: Get report by ID
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Report retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SecurityReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...